pyproject.toml
[tool.poetry]
name = "bic-indexer"
version = "0.1.0"
description = ""
authors = ["Ben <ben@ahnlab.io>"]
readme = "README.md"
packages = [{include = "bic_indexer"}]

[tool.poetry.dependencies]
python = "^3.10"
web3 = "^6.10.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
